requirements.txt
numpy==1.19.5
pandas==1.1.4
pillow==8.1.0
pytorch==1.8.0
pyyaml==5.4.1
torchvision==0.9.0
tqdm==4.56.0
